+++ 2nd-gen/packages/swc/components/divider/index.ts
@@ -9,14 +9,4 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { defineElement } from '@spectrum-web-components/core/element/index.js';
-
-import { Divider } from './Divider.js';
-
 export * from './Divider.js';
-declare global {
-  interface HTMLElementTagNameMap {
-    'swc-divider': Divider;
-  }
-}
-defineElement('swc-divider', Divider);